LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_suppress_compat_events_on_drag_mouse-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Event: Suppress compatibility mouse events on drag

When a pointerdown is canceled, a mouse drag shouldn't fire any compatibility mouse events.

Drag mouse within Target0 & release.
Drag mouse within Target1 & release.
Click Done.
Target0
Target1
Done
The following pointer types were detected: (none).

The following events were logged: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN Suppress compat mouse events on drag
